updated old entities and added category, subcategory and comment entity

+++ src/Rutracker.Server.DataAccessLayer/Interfaces/ITorrentRepository.cs
@@ -4,7 +4,7 @@ using Rutracker.Server.DataAccessLayer.Entities;
 
 namespace Rutracker.Server.DataAccessLayer.Interfaces
 {
-    public interface ITorrentRepository : IRepository<Torrent, long>
+    public interface ITorrentRepository : IRepository<Torrent, int>
     {
         IQueryable<Torrent> Search(string search, long[] forumIds, long? sizeFrom, long? sizeTo);
         IQueryable<Tuple<long, string, int>> GetForums(int count);
